+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 
 ## [Unreleased]
+
+## [0.7.0] - 2016-03-28
 ### Added
 - Activate hovered dropdown item.
 - Add `rotate` dropdown option.
@@ -97,7 +99,8 @@ This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/yuku-t/textcomplete/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/yuku-t/textcomplete/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/yuku-t/textcomplete/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/yuku-t/textcomplete/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/yuku-t/textcomplete/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/yuku-t/textcomplete/compare/v0.4.0...v0.5.0